major revision, input crystals using file, non-orthogonal loops, can go to negative values with edge dislocation, some documentation. Need to verify for screw and bcc structure

+++ src/crystal.cpp
@@ -13,7 +13,7 @@ Crystal::Crystal(){
 }
 
 Crystal::Crystal(int cid,
-                 const Eigen::Matrix<float,3,2>& ctiles,
+                 const Eigen::Matrix<int,3,2>& ctiles,
                  UnitCell* cuc){
   this->_id = cid;
   this->uc = cuc;
